requirements.txt
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
neo4j-driver==1.7.2
neobolt==1.7.4
neotime==1.7.4
pymongo==3.7.2
pytz==2018.9
six==1.12.0
Werkzeug==0.15.1
